requirements.txt
flask==3.0.3
gunicorn==22.0.0
pandas==2.2.2
requests==2.32.3
python-dotenv==1.0.1
clearml==1.16.5
